Renamed subfolders modules/content-items to use camelCase.
