+++ app/modules/contentItems/components/EditableDisplay/index.js
