+++ app/modules/contentItems/components/EditableDisplay/tests/index.test.js
@@ -36,7 +36,7 @@ import type {
   ContentItemsState,
 } from '../../../model';
 
-import * as dummyContentItemData from '../../../lib/test-resources/dummyContentItemData';
+import * as dummyContentItemData from '../../../lib/testResources/dummyContentItemData';
 
 describe(`EditableDisplay`, (): void => {
