+++ app/modules/contentItems/components/EditableDisplay/types/Root.js
