+++ app/modules/contentItems/lib/edit/addChildOrSubItemIdToContext/tests/index.test.js
@@ -13,7 +13,7 @@ import type {
   ParagraphContentItem,
   ContentItemsById,
 } from '../../../../model';
-import * as dummyData from '../../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../../lib/testResources/dummyContentItemData';
 
 import addChildOrSubItemIdToContext from '..';
