+++ app/modules/contentItems/lib/edit/removeChildOrSubItemIdFromContext/tests/index.test.js
@@ -12,7 +12,7 @@ import type {
   ParagraphContentItem,
   ContentItemsById,
 } from '../../../../model';
-import * as dummyData from '../../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../../lib/testResources/dummyContentItemData';
 
 import removeChildOrSubItemIdFromContext from '..';
