+++ app/modules/contentItems/lib/edit/validateChildOrSubItemsInContext/tests/index.test.js
@@ -10,7 +10,7 @@ import type {
   HeadingContentItem,
   ParagraphContentItem,
 } from '../../../../model';
-import * as dummyData from '../../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../../lib/testResources/dummyContentItemData';
 
 import validateChildOrSubItemsInContext from '..';
