+++ app/modules/contentItems/lib/find/allDescendantItems/tests/index.test.js
@@ -6,7 +6,7 @@ import type {
   ParagraphContentItem,
   ContentItemsById,
 } from '../../../../model';
-import * as dummyData from '../../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../../lib/testResources/dummyContentItemData';
 
 import findAllDescendantItems from '..';
