+++ app/modules/contentItems/lib/find/allSiblingItems/tests/index.test.js
@@ -8,7 +8,7 @@ import type {
   HeadingContentItem,
   ParagraphContentItem,
 } from '../../../../model';
-import * as dummyData from '../../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../../lib/testResources/dummyContentItemData';
 
 import findAllSiblingItems from '..';
