+++ app/modules/contentItems/lib/find/furthest/tests/index.test.js
@@ -8,7 +8,7 @@ import type {
   ParagraphContentItem,
   ContentItemsById,
 } from '../../../../model';
-import * as dummyData from '../../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../../lib/testResources/dummyContentItemData';
 
 import find from '../..';
 import findFurthest from '..';
