+++ app/modules/contentItems/lib/testResources/dummyContentItemData.js
