+++ app/modules/contentItems/reducer/tests/moveInState.test.js
@@ -14,7 +14,7 @@ import type {
   ParagraphContentItem,
   ContentItemsState,
 } from '../../model';
-import * as dummyData from '../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../lib/testResources/dummyContentItemData';
 import edit from '../../lib/edit';
 
 import reducer from '..';
