+++ app/modules/contentItems/reducer/tests/removeFromState.test.js
@@ -12,7 +12,7 @@ import type {
   ParagraphContentItem,
   ContentItemsState,
 } from '../../model';
-import * as dummyData from '../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../lib/testResources/dummyContentItemData';
 
 import reducer from '../../reducer';
