+++ app/modules/contentItems/saga/task/tests/add.test.js
@@ -13,7 +13,7 @@ import type {
   ParagraphContentItem,
   ContentItemsById,
 } from '../../../model';
-import * as dummyContentItemData from '../../../lib/test-resources/dummyContentItemData';
+import * as dummyContentItemData from '../../../lib/testResources/dummyContentItemData';
 
 import addSaga from '../add';
