+++ app/modules/contentItems/saga/task/tests/edit.test.js
@@ -14,7 +14,7 @@ import type {
   ContentItemsState,
 } from '../../../model';
 import { getById } from '../../../selectors';
-import * as dummyData from '../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../lib/testResources/dummyContentItemData';
 
 import editSaga from '../edit';
