+++ app/modules/contentItems/saga/task/tests/indent.test.js
@@ -13,7 +13,7 @@ import type {
   ContentItemsById,
   ContentItemsState,
 } from '../../../model';
-import * as dummyData from '../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../lib/testResources/dummyContentItemData';
 
 import indentSaga from '../indent';
