+++ app/modules/contentItems/saga/task/tests/removeAndTogglePreviousItem.test.js
@@ -10,7 +10,7 @@ import type {
   ParagraphContentItem,
   ContentItemsById,
 } from '../../../model';
-import * as dummyData from '../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../lib/testResources/dummyContentItemData';
 
 import removeAndTogglePreviousItemSaga from '../removeAndTogglePreviousItem';
