+++ app/modules/contentItems/saga/task/tests/reverseIndent.test.js
@@ -12,7 +12,7 @@ import type {
   ContentItemsById,
   ContentItemsState,
 } from '../../../model';
-import * as dummyData from '../../../lib/test-resources/dummyContentItemData';
+import * as dummyData from '../../../lib/testResources/dummyContentItemData';
 
 import reverseIndentSaga from '../reverseIndent';
